Update code and workflows
