+++ .github/workflows/pull-request.yml
@@ -1,30 +1,66 @@
-name: CI
+name: Analyze Pull Request
 
 on:
   pull_request:
-    branches: [ main ]
-  push:
-    branches: [ main ]
+    paths-ignore:
+      - '*.md'
+      - 'manifest.json'
+      - '.github/**/*'
+      - 'Dockerfile'
+      - 'Dockerfile.migration'
+      - 'docker-compose.ci.yml'
+      - 'tools/**'
+      - 'screenshots/**'
+      - '.gitignore'
+      - '.editorconfig'
+
+concurrency:
+  # github.workflow: name of the workflow
+  # github.event.pull_request.number || github.ref: pull request number or branch name if not a pull request
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+
+  # Cancel in-progress runs when a new workflow with the same group name is triggered
+  cancel-in-progress: true
 
 jobs:
-  build:
+  test:
     runs-on: ubuntu-latest
+
+    services:
+      db:
+        image: postgis/postgis:14-3.3
+        env:
+          POSTGRES_USER: local
+          POSTGRES_PASSWORD: local
+          POSTGRES_DB: local
+        ports:
+          - 5432:5432
+
     steps:
       - name: Checkout source
-        uses: actions/checkout@v4
+        uses: actions/checkout@v3
 
       - name: Set up JDK
-        uses: actions/setup-java@v4
+        uses: actions/setup-java@v3
         with:
           java-version: '17'
           distribution: 'temurin'
 
+      - name: Run Redis cluster
+        uses: vishnudxb/redis-cluster@1.0.5
+        with:
+          master1-port: 30001
+          master2-port: 30002
+          master3-port: 30003
+          slave1-port: 30004
+          slave2-port: 30005
+          slave3-port: 30006
+
       - name: Setup Gradle
-        uses: gradle/gradle-build-action@v3
+        uses: gradle/gradle-build-action@v2
 
-      - name: Run tests
-        run: ./gradlew test --fail-fast
+      - name: Gradle Clean
+        run: ./gradlew clean
 
-      - name: Publish to Maven repository
-        if: github.ref == 'refs/heads/main'
-        run: ./gradlew publish
+      - name: Gradle Build & Test
+        run: ./gradlew build
