+++ src/test/kotlin/com/c0x12c/featureflag/service/FeatureFlagIntegrationTest.kt
@@ -1,73 +1,29 @@
 package com.c0x12c.featureflag.service
 
-import com.c0x12c.featureflag.cache.RedisCache
-import com.c0x12c.featureflag.repository.FeatureFlagRepository
 import com.c0x12c.featureflag.exception.FeatureFlagNotFoundError
-import org.jetbrains.exposed.sql.Database
-import org.jetbrains.exposed.sql.transactions.transaction
+import com.c0x12c.featureflag.service.utils.TestUtils
+import com.c0x12c.featureflag.service.utils.TestUtils.redisCache
+import com.c0x12c.featureflag.service.utils.TestUtils.service
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import redis.clients.jedis.HostAndPort
-import redis.clients.jedis.JedisCluster
-import java.nio.file.Files
-import java.nio.file.Paths
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 class FeatureFlagIntegrationTest {
 
   companion object {
-    private lateinit var database: Database
-    private lateinit var repository: FeatureFlagRepository
-    private lateinit var redisCache: RedisCache
-    private lateinit var service: FeatureFlagService
-
     @BeforeAll
     @JvmStatic
     fun setup() {
-      setupDatabase()
-      setupRedisCache()
-
-      repository = FeatureFlagRepository(database)
-      service = FeatureFlagService(repository, redisCache)
-    }
-
-    private fun setupDatabase() {
-      database = Database.connect(
-        url = "jdbc:postgresql://localhost:5432/local",
-        driver = "org.postgresql.Driver",
-        user = "local",
-        password = "local"
-      )
-
-      val setupSql = Files.readString(Paths.get("src/test/resources/setup.sql"))
-      transaction(database) {
-        exec(setupSql)
-      }
-    }
-
-    private fun setupRedisCache() {
-      val redisNodes = setOf(
-        HostAndPort("localhost", 30001),
-        HostAndPort("localhost", 30002),
-        HostAndPort("localhost", 30003),
-        HostAndPort("localhost", 30004),
-        HostAndPort("localhost", 30005),
-        HostAndPort("localhost", 30006)
-      )
-      val jedisCluster = JedisCluster(redisNodes)
-      redisCache = RedisCache(jedisCluster, "test")
+      TestUtils.setup()
     }
   }
 
   @BeforeEach
   fun clearData() {
-    transaction(database) {
-      exec("TRUNCATE TABLE feature_flags")
-    }
-
+    TestUtils.clearData()
     redisCache.clearAll()
   }
